+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## Unreleased
+
+- [Fixed] Handle `SIGINT` and `SIGQUIT` signals to gracefully stop recording.
+
 ## v0.0.8
 
 - [Added] Add `WaitUntilDone`, which is just a shortcut for